Renamed FilterBar to more accurate ClearFiltersButton
