+++ realestate.html
@@ -157,6 +157,9 @@
     automagic.addCheckboxFacet("typeFacet", "type", "string");
     // checkbox facet for collection field tags
     automagic.addCheckboxFacet("tagsFacet", "tags", "collection");
+    // filter header & footer options 
+    automagic.addClearFiltersButton("facetHeader");
+    automagic.addClearFiltersButton("facetFooter");    
 </script>
 <style>
