+++ realestate.local.html
@@ -157,9 +157,9 @@
     automagic.addCheckboxFacet("typeFacet", "type", "string");
     // checkbox facet for collection field tags
     automagic.addCheckboxFacet("tagsFacet", "tags", "collection");
-    // header & footer for navigation
-    automagic.addFilterBar("facetHeader");
-    automagic.addFilterBar("facetFooter");
+    // filter header & footer options 
+    automagic.addClearFiltersButton("facetHeader");
+    automagic.addClearFiltersButton("facetFooter");
 </script>
 <style>
