+++ src/AzSearch.tsx
@@ -10,18 +10,18 @@ import { ResultsContainer, OwnProps as ResultsOwnProps }from "./containers/Resul
 import { CheckboxFacetContainer, OwnProps as CheckboxOwnProps } from "./containers/CheckboxFacetContainer";
 import { RangeFacetContainer, OwnProps as RangeOwnProps } from "./containers/RangeFacetContainer";
 import { PagerContainer, OwnProps as PagerOwnProps } from "./containers/PagerContainer";
-import { FilterBarContainer, OwnProps as FilterBarOwnProps } from "./containers/FilterBarContainer";
+import { ClearFiltersButtonContainer, OwnProps as FilterBarOwnProps } from "./containers/ClearFiltersButtonContainer";
 import SearchBox from "./components/SearchBox";
 import CheckboxFacet from "./components/CheckboxFacet";
 import RangeFacet from "./components/CheckboxFacet";
-import FilterBar from "./components/FilterBar";
+import ClearFiltersButton from "./components/ClearFiltersButton";
 import Results from "./components/Results";
 import Pager from "./components/Pager";
 
 import "rc-slider/assets/index.css";
 
-let Components = { SearchBox, CheckboxFacet, Results, FilterBar };
-let Containers = { CheckboxFacetContainer, ResultsContainer, SearchBoxContainer, FilterBarContainer };
+let Components = { SearchBox, CheckboxFacet, Results, ClearFiltersButton };
+let Containers = { CheckboxFacetContainer, ResultsContainer, SearchBoxContainer, ClearFiltersButtonContainer };
 
 class Automagic {
     public store: AzSearchStore;
@@ -82,10 +82,10 @@ class Automagic {
         );
     }
 
-    public addFilterBar(htmlId: string) {
+    public addClearFiltersButton(htmlId: string) {
         render(
             <Provider store={this.store.store}>
-                <FilterBarContainer/>
+                <ClearFiltersButtonContainer/>
             </Provider>,
             document.getElementById(htmlId)
         );
